Move data types to seperate file
